+++ lib/authorization/endpoints.js
@@ -2,6 +2,7 @@
 
 var _ = require('lodash');
 var express = require('express');
+var pick = require('../utils/pick');
 
 var consts = require('./../constants');
 
@@ -37,7 +38,7 @@ function init (env, authorization) {
   });
   endpoints.get('/subjects', authorization.isPermitted('admin:api:subjects:read'), function getSubjects (req, res) {
     res.json(authorization.storage.subjects.map(function eachSubject (subject) {
-      return _.pick(subject, ['_id', 'name', 'accessToken', 'roles']);
+      return pick(subject, ['_id', 'name', 'accessToken', 'roles']);
     }));
   });
