+++ lib/client/careportal.js
@@ -4,21 +4,9 @@ var parse_duration = require('parse-duration'); // https://www.npmjs.com/package
 var times = require('../times');
 var consts = require('../constants');
 var Storages = require('js-storage');
+var pick = require('../utils/pick');
 
-/**
- * Helper function to pick specific properties from an object
- * @param {Object} obj - Source object
- * @param {string[]} keys - Array of keys to pick
- * @returns {Object} - New object with only the specified keys
- */
-function pick(obj, keys) {
-  return keys.reduce((result, key) => {
-    if (Object.prototype.hasOwnProperty.call(obj, key)) {
-      result[key] = obj[key];
-    }
-    return result;
-  }, {});
-}
+
 
 /**
  * @typedef {Object} Reason
