+++ lib/data/treatmenttocurve.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var consts = require('../constants');
+var pick = require('../utils/pick');
 
 const MAX_BG_MMOL = 22;
 const MAX_BG_MGDL = MAX_BG_MMOL * consts.MMOL_TO_MGDL;
@@ -70,7 +71,7 @@ module.exports = function fitTreatmentsToBGCurve (ddata, env, ctx) {
       }
     } else if (treatment.glucose) {
       //no units, assume everything is the same
-      //console.warn('found a glucose value without any units, maybe from an old version?', _.pick(treatment, '_id', 'created_at', 'enteredBy'));
+      //console.warn('found a glucose value without any units, maybe from an old version?', pick(treatment, ['_id', 'created_at', 'enteredBy']));
       var units = settings.units === 'mmol' ? 'mmol' : 'mgdl';
 
       treatment[units] = settings.units === 'mmol' ? Math.min(Number(treatment.glucose), MAX_BG_MMOL) : Math.min(Number(treatment.glucose), MAX_BG_MGDL);
