+++ lib/plugins/upbat.js
@@ -2,6 +2,7 @@
 
 var _ = require('lodash');
 var times = require('../times');
+var pick = require('../utils/pick');
 
 function init(ctx) {
   var translate = ctx.language.translate;
@@ -116,7 +117,7 @@ function init(ctx) {
     recentData?.forEach(function eachRecentStatus (status) {
       analyzeStatus(status);
       var device = getDevice(status);
-      device.statuses.push(_.pick(status, ['uploader', 'created_at', 'mills', '_id']));
+      device.statuses.push(pick(status, ['uploader', 'created_at', 'mills', '_id']));
     });
 
     var recentLowests = [ ];
