+++ lib/sandbox.js
@@ -3,6 +3,7 @@
 var _ = require('lodash');
 var units = require('./units')();
 var times = require('./times');
+var pick = require('./utils/pick');
 
 function init () {
   var sbx = {};
@@ -30,7 +31,7 @@ function init () {
    * @returns  {{notification}}
    */
   function safeNotifications (ctx) {
-    return _.pick(ctx.notifications, ['requestNotify', 'requestSnooze', 'requestClear']);
+    return pick(ctx.notifications, ['requestNotify', 'requestSnooze', 'requestClear']);
   }
 
   /**
