+++ lib/server/booterror.js
@@ -3,6 +3,7 @@
 const express = require('express');
 const path = require('path');
 var _ = require('lodash');
+var pick = require('../utils/pick');
 
 function bootError(env, ctx) {
 
@@ -23,7 +24,7 @@ function bootError(env, ctx) {
       if (typeof obj.err === 'string' || obj.err instanceof String) {
         message = obj.err;
       } else {
-        message = JSON.stringify(_.pick(obj.err, Object.getOwnPropertyNames(obj.err)));
+        message = JSON.stringify(pick(obj.err, Object.getOwnPropertyNames(obj.err)));
       }
       return '<dt><b>' + obj.desc + '</b></dt><dd>' + message.replace(/\\n/g, '<br/>') + '</dd>';
     }).join(' ');
