+++ lib/server/pebble.js
@@ -132,7 +132,6 @@ function prepareCals (req, sbx) {
   var data = sbx.data;
   if (req.rawbg && data.cals && data.cals.length > 0) {
     return reverseAndSlice(data.cals, req).map(function transformCal (cal) {
-      // Use object destructuring and property shorthand instead of _.pick
       const { slope, intercept, scale } = cal;
       return { slope, intercept, scale };
     });
