+++ lib/utils.js
@@ -61,6 +61,8 @@ function init(ctx) {
     return moment(datestring + ' ' + timestring, 'YYYY-MM-D HH:mm');
   };
 
+  utils.pick = require('./utils/pick');
+
   /**
    * Extract device name from device string
    * @param {string} device - The device string to parse
